Remove news

+++ flaskr/blog/context_processors.py
@@ -51,10 +51,14 @@ def utility_processor():
 def utility_processor_remove_news():
     def remove_news(id):
 
-        db = get_db
-        db.execute('DELETE FROM news WHERE id = ?', (id,))
-        db.commit()
-        return redirect(url_for('blog.index'))
+        if (g.user['is_admin']):
+
+            db = get_db
+            db.execute('DELETE FROM news WHERE id = ?', (id,))
+            db.commit()
+            flash('Removed news with id = {0}'.format(id))
+            return redirect(url_for('blog.index'))
+        flash('Failed to remove news with id = {0}'.format(id))
     return dict(remove_news=remove_news)
 
 #@bp.context_processor
